Added a feature for the user to see the location of the weather currently being shown.
The Location is shown using the apps that give access to the maps on the phone.  For example, Google Maps. But if there are more than one app to show the location of the Zip Code on the map it will prompt the user to choose the app to see the location.